remove redundant factory classes and replace with companion objects for improved simplicity
