+++ src/main/kotlin/org/jetbrains/research/testspark/testmanager/template/TestAnalyzer.kt
@@ -1,5 +1,9 @@
 package org.jetbrains.research.testspark.testmanager.template
 
+import org.jetbrains.research.testspark.core.test.SupportedLanguage
+import org.jetbrains.research.testspark.testmanager.java.JavaTestAnalyzer
+import org.jetbrains.research.testspark.testmanager.kotlin.KotlinTestAnalyzer
+
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.vfs.VirtualFile
 import org.jetbrains.research.testspark.core.test.data.TestSample
@@ -8,6 +12,14 @@ import org.jetbrains.research.testspark.core.test.data.TestSample
  * Interface for retrieving information from test class code.
  */
 interface TestAnalyzer {
+    companion object {
+        fun create(language: SupportedLanguage): TestAnalyzer =
+            when (language) {
+                SupportedLanguage.Kotlin -> KotlinTestAnalyzer
+                SupportedLanguage.Java -> JavaTestAnalyzer
+            }
+    }
+
     /**
      * Extracts the code of the first test method found in the given class code.
      *
