--- src/main/kotlin/org/jetbrains/research/testspark/tools/llm/generation/TestBodyPrinterFactory.kt
@@ -1,14 +1,0 @@
-package org.jetbrains.research.testspark.tools.llm.generation
-
-import org.jetbrains.research.testspark.core.test.SupportedLanguage
-import org.jetbrains.research.testspark.core.test.TestBodyPrinter
-import org.jetbrains.research.testspark.core.test.java.JavaTestBodyPrinter
-import org.jetbrains.research.testspark.core.test.kotlin.KotlinTestBodyPrinter
-
-object TestBodyPrinterFactory {
-    fun create(language: SupportedLanguage): TestBodyPrinter =
-        when (language) {
-            SupportedLanguage.Kotlin -> KotlinTestBodyPrinter()
-            SupportedLanguage.Java -> JavaTestBodyPrinter()
-        }
-}
